#259 [refact] project의 필터링 검색 기능을 search 클래스로 따로 분리하여 리팩토링한다.
